+++ typescript/infra/config/environments/test/middleware/liquidity-layer/addresses.json
@@ -1,11 +1,11 @@
 {
   "fuji": {
-    "circleBridgeAdapter": "0xb54AD7AE42B7c505100594365CdBC4b28Ef51FE6",
+    "circleBridgeAdapter": "0x17EB33454AAEF8E91510540a0ebF4a8213dd740D",
     "portalAdapter": "0xe589a05be1304b43A6FEb9c5D6a6EEEA35656271",
     "router": "0x3428e12EfDb2446c1E7feC3f1CED099A8a7cD541"
   },
   "goerli": {
-    "circleBridgeAdapter": "0xb54AD7AE42B7c505100594365CdBC4b28Ef51FE6",
+    "circleBridgeAdapter": "0x17EB33454AAEF8E91510540a0ebF4a8213dd740D",
     "portalAdapter": "0xe589a05be1304b43A6FEb9c5D6a6EEEA35656271",
     "router": "0x3428e12EfDb2446c1E7feC3f1CED099A8a7cD541"
   },
