+++ typescript/infra/config/environments/test/middleware/liquidity-layer/verification.json
@@ -17,6 +17,18 @@
       "address": "0xe589a05be1304b43A6FEb9c5D6a6EEEA35656271",
       "isProxy": false,
       "constructorArguments": ""
+    },
+    {
+      "name": "CircleBridgeAdapter",
+      "address": "0x54FCA26E5FF828847D8caF471e44cD5727C73B0d",
+      "isProxy": false,
+      "constructorArguments": ""
+    },
+    {
+      "name": "CircleBridgeAdapter",
+      "address": "0x17EB33454AAEF8E91510540a0ebF4a8213dd740D",
+      "isProxy": false,
+      "constructorArguments": ""
     }
   ],
   "goerli": [
@@ -35,6 +47,18 @@
     {
       "name": "PortalAdapter",
       "address": "0xe589a05be1304b43A6FEb9c5D6a6EEEA35656271",
+      "isProxy": false,
+      "constructorArguments": ""
+    },
+    {
+      "name": "CircleBridgeAdapter",
+      "address": "0x54FCA26E5FF828847D8caF471e44cD5727C73B0d",
+      "isProxy": false,
+      "constructorArguments": ""
+    },
+    {
+      "name": "CircleBridgeAdapter",
+      "address": "0x17EB33454AAEF8E91510540a0ebF4a8213dd740D",
       "isProxy": false,
       "constructorArguments": ""
     }
